feat: GraphQL subscription (SSE) and EventSource integration
