+++ packages/bff/package.json
@@ -53,7 +53,6 @@
     "fastify-plugin": "^4.5.1",
     "graphql": "^16.8.1",
     "graphql-http": "^1.22.0",
-    "graphql-sse": "^2.5.3",
     "graphql-tools": "^9.0.1",
     "ioredis": "^5.4.0",
     "jsonwebtoken": "^9.0.2",
